# planner-api/openapi.yaml
openapi: "3.0.3"

info:
  title: "Savings Planner API"
  description: "HTTP endpoints docs"
  version: "1.0"

paths:
  "/api/auth/users":
    post:
      summary: "Creates new user account"
      requestBody:
        description: "User data"
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/RegistrationRequest"
      responses:
        "201":
          description: "User has been created"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/AccessTokenResponse"
        "400":
          description: "Invalid request body"
        "409":
          description: "Email or nick already taken"

  "/api/auth/access-token":
    post:
      summary: "Gets new access token"
      requestBody:
        description: "User credentials"
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: "Login successful"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/AccessTokenResponse"
        "400":
          description: "Invalid request body"
        "401":
          description: "Wrong login or password"

  "/api/balance":
    put:
      summary: "Update user balance"
      requestBody:
        description: "New balance"
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/BalanceUpdateRequest"
      responses:
        "204":
          description: "Balance has been updated"
        "401":
          description: "User unauthorized"

    get:
      summary: "Get actual user balance"
      responses:
        "200":
          description: "Got user balance"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/BalanceListResponse"
        "401":
          description: "User unauthorized"

  "/api/events":
    post:
      summary: "Creates new one-time event"
      requestBody:
        description: "Event description"
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/EventInfo"
      responses:
        "201":
          description: "Event has been created"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/EventData"
        "400":
          description: "Invalid request body"
        "401":
          description: "User unauthorized"
    get:
      summary: "Gets list of user's one-time events"
      responses:
        "200":
          description: "Got list of events"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/EventListResponse"
        "401":
          description: "User unauthorized"


  "/api/events/{id}":
    put:
      summary: "Updates one-time event"
      requestBody:
        description: "Event description"
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/EventInfo"
      responses:
        "204":
          description: "Event has been updated"
        "400":
          description: "Invalid request body"
        "401":
          description: "User unauthorized"

    delete:
      summary: "Deletes one-time event"
      responses:
        "204":
          description: "Event successfully deleted"
        "401":
          description: "User unauthorized"
        "404":
          description: "Event with that ID does not exist"

  "/api/goals":
    post:
      summary: "Create goal"
      requestBody:
        description: "Goal data"
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/GoalCreateRequest"
      responses:
        "201":
          description: "Goal successfully created"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/GoalData"
        "401":
          description: "User unauthorized"
    get:
      summary: "Gets list of user's goals"
      responses:
        "200":
          description: "Got list of user's goals"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/GoalListResponse"
        "401":
          description: "User unauthorized"
    patch:
      summary: "Update goals priorities"
      requestBody:
        description: "IDs and new priorities of goals to update"
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/GoalPriorityUpdate"
      responses:
        "204":
          description: "Update was successful"
        "401":
          description: "User unauthorized"
        "404":
          description: "Goal with the given ID not found among user's goals"

  "/api/goals/{id}":
    delete:
      summary: "Deletes goal"
      responses:
        "204":
          description: "Goal successfully deleted"
        "401":
          description: "User unauthorized"
        "404":
          description: "Goal with that ID does not exist"

components:
  schemas:
    RegistrationRequest:
      type: "object"
      properties:
        nick:
          type: "string"
          pattern: ^\w{3,16}$
        email:
          type: "string"
          format: "email"
        password:
          type: "string"
          format: "password"
          minLength: 7
          maxLength: 100

    LoginRequest:
      type: "object"
      properties:
        login:
          type: "string"
          format: "nick or email"
        password:
          type: "string"
          format: "password"
          minLength: 7
          maxLength: 100

    AccessTokenResponse:
      type: "object"
      properties:
        accessToken:
          type: "string"
          format: "encoded access token"
          example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ1c2VyMiIsImV4cCI6MTY0OTk2MTYzOCwidXNlcklkIjoxfQ.rWcKHXk8Q-QWchLEEBRlxSS3owbbQa5azSM4ag5ogpY"

    AccessToken:
      type: "object"
      description: "Decoded token structure"
      properties:
        header:
          type: "object"
          properties:
            alg:
              type: "string"
              default: "HS256"
            typ:
              type: "string"
              default: "JWT"
        payload:
          type: "object"
          properties:
            sub:
              type: "string"
              format: "nick"
              description: "Owner of the token"
            exp:
              type: "integer"
              format: "int64"
              description: "Expiration (Unix epoch second)"
            userId:
              type: "integer"
              format: "int32"
              description: "Owner's ID"

    BalanceUpdateRequest:
      type: "object"
      properties:
        balance:
          type: "integer"
          format: "positive or zero account balance"
          example: 10000

    BalanceListResponse:
      type: "object"
      properties:
        balance:
          type: "integer"
          format: "positive or zero account balance"
          example: 0

    EventInfo:
      type: "object"
      properties:
        title:
          type: "string"
          format: "event name"
          minLength: 1
          maxLength: 200
          example: "Repayment"
        amount:
          type: "integer"
          format: "positive or negative money gain"
          example: -205
        timestamp:
          type: "integer"
          example: "100000"

    EventListResponse:
      type: "object"
      properties:
        list:
          type: "array"
          items:
            type: "object"
            properties:
              id:
                type: "integer"
                format: "id of an object"
                example: 1
              title:
                type: "string"
                format: "event name"
                minLength: 1
                maxLength: 200
                example: "Repayment"
              amount:
                type: "integer"
                format: "positive or negative money gain"
                example: 201
              timestamp:
                type: "integer"
                example: "100000"

    EventData:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "id of an object"
          example: 1
        title:
          type: "string"
          format: "event name"
          minLength: 1
          maxLength: 200
          example: "Repayment"
        amount:
          type: "integer"
          format: "positive or negative money gain"
          example: 201
        timestamp:
          type: "integer"
          example: "100000"

    GoalListResponse:
      type: "object"
      properties:
        list:
          type: "array"
          items:
            type: "object"
            properties:
              id:
                type: "integer"
                format: "id of an object"
                example: 1
              title:
                type: "string"
                format: "goal name"
                minLength: 1
                maxLength: 200
                example: "Car"
              amount:
                type: "integer"
                format: "cost of fulfilling the goal"
                example: 201

    GoalCreateRequest:
      type: "object"
      properties:
        list:
          type: "array"
          items:
            type: "object"
            properties:
              title:
                type: "string"
                format: "goal name"
                minLength: 1
                maxLength: 200
                example: "Car"
              amount:
                type: "integer"
                format: "cost of fulfilling the goal"
                example: 201

    GoalPriorityUpdate:
      type: "object"
      properties:
        newPriorities:
          type: "array"
          items:
            type: "object"
            properties:
              id:
                type: "integer"
              newPriority:
                type: "integer"

    GoalData:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "id of an object"
          example: 1
        title:
          type: "string"
          format: "goal name"
          minLength: 1
          maxLength: 200
          example: "Car"
        amount:
          type: "integer"
          format: "cost of fulfilling the goal"
          example: 201